refactor: 온보딩 API를 /users/me/medicines 구조로 전환
