+++ src/main/java/com/mediforme/mediforme/dto/request/OnboardingRequestDto.java
@@ -8,11 +8,9 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 public class OnboardingRequestDto {
-    private String userLoginId;   // 로그인한 사용자 ID (User 엔티티 참조)
     private String itemName;      // 약 이름 (Medicine)
     private Long mealCd;          // 식사 코드 (before/after meal)
     private Long timeCd;          // 복용 시간 코드 (morning/lunch/dinner)
     private Long daysOfWeekCd;    // 요일 코드
     private String dosage;        // 복용량
-    private Boolean isAlarm;      // 알람 여부
 }
