+++ src/main/java/com/mediforme/mediforme/service/OnboardingService.java
@@ -8,7 +8,7 @@ import java.io.IOException;
 
 public interface OnboardingService {
     // 약 등록, 조회, 삭제
-    OnboardingResponseDto saveMedicineInfo(OnboardingRequestDto requestDto) throws IOException, ParseException;
+    OnboardingResponseDto saveMedicineInfo(Long userId, OnboardingRequestDto requestDto) throws IOException, ParseException;
     OnboardingResponseDto getUserMedicines(Long userId);
     void deleteUserMedicine(Long userMedicineId, Long userId);
 }
